Comentado y con printf
