+++ src/queue.h
@@ -1,11 +1,12 @@
-// #ifndef QUEUE_H
-// #define QUEUE_H
+#ifndef QUEUE_H
+#define QUEUE_H
 
-// #include "simulation.h"
+#include "simulation.h"
+#include <stdio.h>
+#include <stdlib.h>
 
-// queue_t *createQueue(unsigned capacity);
-// void enqueue(queue_t *queue, proc_info_t process);
-// proc_info_t dequeue(queue_t *queue);
+queue_t *createQueue(unsigned capacity) ;
+void enqueue(queue_t *queue, proc_info_t process);
+proc_info_t dequeue(queue_t *queue);
 
-
-// #endif
+#endif
